Refactor FixedHeader and FixedFooter to be a single component with higher-order component shells
